Fixing | Users permissions

+++ users/resolvers.py
@@ -2,6 +2,8 @@ from users.models import User
 from django.contrib.auth.hashers import make_password
 import uuid
 
+from graphtest.errors import *
+
 def resolve_create_user(user_input):
     hashed_password = make_password(user_input.password)
     user_instance = User(
@@ -14,22 +16,28 @@ def resolve_create_user(user_input):
 
     return user_instance
 
-def resolve_update_user(user_input, kwargs):
-    user_instance = User.objects.get(pk=kwargs.get('id'))
-    if user_input.username:
-      user_instance.username = user_input.username
-    if user_input.email:
-      user_instance.email = user_input.email
-    if user_input.bio:
-      user_instance.bio = user_input.bio
-    if user_input.password:
-      user_instance.password = make_password(user_input.password)
-    user_instance.save()
+def resolve_update_user(logged_user_id, user_input, kwargs):
+    user_id = kwargs.get('id')
+    if logged_user_id == user_id:
+        user_instance = User.objects.get(pk=user_id)
+        if user_input.username:
+          user_instance.username = user_input.username
+        if user_input.email:
+          user_instance.email = user_input.email
+        if user_input.bio:
+          user_instance.bio = user_input.bio
+        if user_input.password:
+          user_instance.password = make_password(user_input.password)
+        user_instance.save()
 
-    return user_instance
+        return user_instance
+    has_permissions()
 
-def resolve_delete_user(kwargs):
-    user = User.objects.get(pk=kwargs.get('id'))
-    user.delete()
+def resolve_delete_user(logged_user_id, kwargs):
+    user_id = kwargs.get('id')
+    if logged_user_id ==  user_id:
+        user = User.objects.get(pk=user_id)
+        user.delete()
 
-    return True
+        return True
+    has_permissions()
